requirements.txt
appdirs==1.4.4
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astroid==2.5.1
autopep8==1.5.5
backcall==0.2.0
certifi==2020.12.5
colorama==0.4.4
decorator==5.0.5
distlib==0.3.1
Django @ file:///tmp/build/80754af9/django_1613762073373/work
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Faker==7.0.1
filelock==3.0.12
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycodestyle==2.6.0
Pygments==2.8.1
PyJWT==2.0.1
pylint==2.7.2
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
virtualenv==20.4.2
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1
